pyproject.toml
[project]
name = "comosoco"
version = "1.2026.0"
description = "Template environment for the COGS 50.09 course"
authors = [{ name = "Sean Dae Houlihan", email = "dae.houlihan@dartmouth.edu" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    ### i/o ###
    "python-dotenv>=1",
    "toml",
    "pyyaml",
    ### scientific computing ###
    "numpy>=2.0.2",
    "optype[numpy]",
    "pandas>=2.3.3",
    "xarray>=2025.12.0",
    "scipy>=1.16.3",
    ### memo ###
    "jax>=0.8.2",
    "memo-lang>=1.2.7",
    ### plotting ###
    "matplotlib>=3.10.0",
    "seaborn>=0.13.0",
    ### interactive ui ###
    "marimo>=0.18.4",
    "anywidget>=0.9.0",
    "altair",
    "plotly<6.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ty>=0.0.9",
    "pandas-stubs",
    "scipy-stubs",
]

[project.urls]
Homepage = "https://comosoco.daeh.info/"
Repository = "https://github.com/daeh/comosoco-env"
Documentation = "https://github.com/daeh/comosoco-env?tab=readme-ov-file#readme"
Issues = "https://github.com/daeh/comosoco-env/issues"

[tool.pyright]
pythonVersion = "3.14"
# reportGeneralTypeIssues = "error"
reportMissingImports = "error"
# reportMissingTypeStubs = true
reportUnusedImport = "none"
reportUnusedClass = "error"
reportUnusedFunction = "warning"
# reportUnusedVariable = "information"
# reportUnboundVariable = "information"
### memo ###
reportUnusedVariable = "none"
reportUnboundVariable = "none"
reportGeneralTypeIssues = "none"
reportUndefinedVariable = "none"
reportInvalidTypeForm = "none"
reportAttributeAccessIssue = "none"
reportIndexIssue = "none"
reportCallissue = "none"
reportOperatorIssue = "none"
# reportRedeclaration = "warning"

[tool.ty.src]
# Exclude student code directory from ty type checking (memo-lang DSL)
exclude = ["comosoco/"]
